moves all groups components from GroupManagement dir to its own Groups dir
