+++ src/login/components/GroupManagement/Groups/GroupParent.js
@@ -1,28 +1,28 @@
 import React, { Component, Fragment } from "react";
 import PropTypes from "prop-types";
-import InjectIntl from "../../translation/InjectIntl_HOC_factory";
+import InjectIntl from "../../../translation/InjectIntl_HOC_factory";
 import GroupList from "./GroupList";
-import EditMode from "./EditMode/EditMode";
+import EditGroup from "./EditGroup";
 
-class GroupDataPanel extends Component {
+class GroupParent extends Component {
   state = {
-    editMode:
-      JSON.parse(window.localStorage.getItem("persistedEditMode")) || false,
+    editGroup:
+      JSON.parse(window.localStorage.getItem("persistedEditGroup")) || false,
     group: JSON.parse(window.localStorage.getItem("persistedGroup")) || "",
   };
 
   componentDidUpdate() {
-    window.localStorage.setItem("persistedEditMode", this.state.editMode);
+    window.localStorage.setItem("persistedEditGroup", this.state.editGroup);
     window.localStorage.setItem(
       "persistedGroup",
       JSON.stringify(this.state.group)
     );
   }
 
-  toggleViewOrEditMode = (singleGroupData) => {
+  toggleGroupListOrEditGroup = (singleGroupData) => {
     this.setState((prevState) => {
       return {
-        editMode: !prevState.editMode,
+        editGroup: !prevState.editGroup,
         group: singleGroupData,
       };
     });
@@ -31,16 +31,16 @@ class GroupDataPanel extends Component {
   render() {
     return (
       <div className="data-panel">
-        {this.state.editMode ? (
-          <EditMode
+        {this.state.editGroup ? (
+          <EditGroup
             {...this.props}
             group={this.state.group}
-            toggleViewOrEditMode={this.toggleViewOrEditMode}
+            toggleGroupListOrEditGroup={this.toggleGroupListOrEditGroup}
           />
         ) : (
           <GroupList
             {...this.props}
-            toggleViewOrEditMode={this.toggleViewOrEditMode}
+            toggleGroupListOrEditGroup={this.toggleGroupListOrEditGroup}
           />
         )}
       </div>
@@ -48,6 +48,6 @@ class GroupDataPanel extends Component {
   }
 }
 
-GroupDataPanel.propTypes = {};
+GroupParent.propTypes = {};
 
-export default InjectIntl(GroupDataPanel);
+export default InjectIntl(GroupParent);
